+++ src/Products/Five/browser/resource.py
@@ -26,7 +26,6 @@ from zope.publisher.interfaces import NotFound
 from zope.publisher.interfaces.browser import IBrowserPublisher
 from zope.ptresource.ptresource import PageTemplate
 
-from Acquisition import aq_base
 from Products.Five.browser import BrowserView
 
 
@@ -175,7 +174,7 @@ class DirectoryResource(Resource,
 
         # We need to propagate security so that restrictedTraverse() will
         # work
-        if hasattr(aq_base(self), '__roles__'):
+        if hasattr(self, '__roles__'):
             resource.__roles__ = self.__roles__
 
         return resource