+++ src/Products/Five/browser/tests/test_pages.py
@@ -74,6 +74,4 @@ def test_suite():
         ZopeDocFileSuite('pages.txt', package='Products.Five.browser.tests'),
         FunctionalDocFileSuite('pages_ftest.txt',
                                package='Products.Five.browser.tests'),
-        FunctionalDocFileSuite('aqlegacy_ftest.txt',
-                               package='Products.Five.browser.tests'),
     ))